improve handling encryption token, make Challenge Encryption
rename current encryption system to Challenge Encryption because we plan to add basic encryption, less complicated system for the data with low sensitivity

+++ app/Rules/VerifyChallengeEncryptionTokenRule.php
@@ -3,19 +3,12 @@
 namespace App\Rules;
 
 use App\Models\EncryptionToken;
-use App\Services\Encryption\Token\TokenFactory;
+use App\Services\Encryption\Challenge\TokenFactory;
 use Closure;
 use Illuminate\Contracts\Validation\ValidationRule;
 use Illuminate\Http\Request;
-use Symfony\Component\HttpFoundation\Response;
 
-/**
- * HTTP error status codes:
- *      400 - invalid token is submitted. User flow - new token needs to be created
- *      404 - token is not created at all (for this user and session)
- *      419 - token is expired, and user needs to create a new one
- */
-class VerifyEncryptionTokenRule implements ValidationRule
+class VerifyChallengeEncryptionTokenRule implements ValidationRule
 {
     /**
      * VerifyEncryptionTokenRule constructor.
@@ -31,39 +24,33 @@ class VerifyEncryptionTokenRule implements ValidationRule
      */
     public function validate(string $attribute, mixed $value, Closure $fail): void
     {
-        $encryptedToken = $this->getEncryptedToken();
+        if (!($encryptedToken = $this->getEncryptedToken())) {
+            $fail(__('encryption.token.not_found'));
+            return;
+        }
 
         if (!TokenFactory::verify($encryptedToken, $value)) {
             // invalid token is submitted
-            abort(Response::HTTP_BAD_REQUEST, __('encryption.token.failed'));
+            $fail(__('encryption.token.failed'));
+            return;
         }
 
         if (TokenFactory::isExpired($encryptedToken)) {
             // if token ends his short life, delete it from the DB
             $encryptedToken->delete();
-            abort(419, __('encryption.token.expired'));
+            $fail(__('encryption.token.expired'));
         }
-
-        // if token verification is passed, extend token life
-        $encryptedToken->forceFill([
-            'last_used_at' => now(),
-            'expires_at'   => now()->addMinutes(config('auth.encryption_token.expire')),
-        ])->save();
     }
 
     /**
      * Get the encrypted token from the DB.
      *
-     * @return EncryptionToken
+     * @return EncryptionToken|null
      */
-    protected function getEncryptedToken(): EncryptionToken
+    protected function getEncryptedToken(): ?EncryptionToken
     {
         // reload 'encryptionToken' relation
         $user = $this->request->user()->load('encryptionToken');
-
-        if (empty($user->encryptionToken)) {
-            abort(Response::HTTP_NOT_FOUND, __('encryption.token.not_found'));
-        }
 
         return $user->encryptionToken;
     }
